moving contracts to constant
